all dockered and project rename
