+++ build.gradle
@@ -1,12 +1,12 @@
 plugins {
 	id 'java'
-	id 'war'
 	id 'org.springframework.boot' version '3.3.3'
 	id 'io.spring.dependency-management' version '1.1.6'
 }
 
-group = 'com.nordclan.nikgapon'
-version = '0.0.1-SNAPSHOT'
+group = 'com.nikgapon'
+version = '0.0.2'
+
 
 java {
 	toolchain {
@@ -21,7 +21,7 @@ repositories {
 dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
-	providedRuntime 'org.springframework.boot:spring-boot-starter-tomcat'
+	runtimeOnly  'org.springframework.boot:spring-boot-starter-tomcat'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 	// https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-thymeleaf
