+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'work_practice_1'
+rootProject.name = 'MeetingRoomsReservator'
