+++ src/main/java/com/nikgapon/MeetingRoomsReservator/MeetingRoomsReservatorApplication.java
@@ -1,0 +1,11 @@
+package com.nikgapon.MeetingRoomsReservator;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MeetingRoomsReservatorApplication {
+    public static void main(String[] args){
+        SpringApplication.run(MeetingRoomsReservatorApplication.class, args);
+    }
+}
