+++ src/main/java/com/nikgapon/MeetingRoomsReservator/controller/MainPageController.java
@@ -1,0 +1,16 @@
+package com.nikgapon.MeetingRoomsReservator.controller;
+
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+public class MainPageController {
+    @RequestMapping("/")
+    public String mainpage(Model model){
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        model.addAttribute("name", username);
+        return "mainpage";
+    }
+}
