+++ src/main/java/com/nikgapon/MeetingRoomsReservator/repository/MeetingRoomRepository.java
@@ -1,0 +1,7 @@
+package com.nikgapon.MeetingRoomsReservator.repository;
+
+import com.nikgapon.MeetingRoomsReservator.model.MeetingRoomEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MeetingRoomRepository extends JpaRepository<MeetingRoomEntity, Long> {
+}
