--- src/main/java/com/nordclan/nikgapon/work_practice_1/PracticeOneApplication.java
@@ -1,12 +1,0 @@
-package com.nordclan.nikgapon.work_practice_1;
-
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PracticeOneApplication {
-    public static void main(String[] args){
-        SpringApplication.run(PracticeOneApplication.class, args);
-    }
-}
