--- src/main/java/com/nordclan/nikgapon/work_practice_1/controller/AuthController.java
@@ -1,23 +1,0 @@
-package com.nordclan.nikgapon.work_practice_1.controller;
-
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class AuthController {
-    @GetMapping("/login")
-    public String login() {
-        return "login";
-    }
-    @GetMapping("/logout")
-    public String logout(HttpServletRequest request) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (authentication != null) {
-            request.getSession().invalidate();
-        }
-        return "redirect:/login";
-    }
-}
