--- src/main/java/com/nordclan/nikgapon/work_practice_1/model/UserRole.java
@@ -1,20 +1,0 @@
-package com.nordclan.nikgapon.work_practice_1.model;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public enum UserRole implements GrantedAuthority {
-    ADMIN,
-    USER;
-
-    private static final String PREFIX = "ROLE_";
-
-    @Override
-    public String getAuthority() {
-        return PREFIX + this.name();
-    }
-
-    public static final class AsString {
-        public static final String ADMIN = PREFIX + "ADMIN";
-        public static final String USER = PREFIX + "USER";
-    }
-}
