+++ src/main/resources/application.properties
@@ -1,10 +1,7 @@
-spring.application.name=work_practice_1
-
-spring.datasource.url=jdbc:postgresql://localhost:5432/NordClanPractice1
+spring.application.name=MeetingRooms-reservator
+spring.datasource.url=jdbc:postgresql://localhost:5432/MeetingRoomsReservator
 spring.datasource.username=postgres
 spring.datasource.password=admin
 spring.datasource.driver-class-name=org.postgresql.Driver
 
-spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
